+++ packages/nesd/lib/ui/settings/settings.g.dart
@@ -87,9 +87,6 @@ _Settings _$SettingsFromJson(Map<String, dynamic> json) => _Settings(
       PixelAspectRatio.auto,
   customPixelAspectRatio:
       (json['customPixelAspectRatio'] as num?)?.toDouble() ?? 1.0,
-  videoFilter:
-      $enumDecodeNullable(_$VideoFilterEnumMap, json['videoFilter']) ??
-      VideoFilter.none,
   videoFilters:
       (json['videoFilters'] as List<dynamic>?)
           ?.map((e) => $enumDecode(_$VideoFilterEnumMap, e))
@@ -131,7 +128,6 @@ Map<String, dynamic> _$SettingsToJson(_Settings instance) => <String, dynamic>{
   'rewind': instance.rewind,
   'pixelAspectRatio': _$PixelAspectRatioEnumMap[instance.pixelAspectRatio]!,
   'customPixelAspectRatio': instance.customPixelAspectRatio,
-  'videoFilter': _$VideoFilterEnumMap[instance.videoFilter]!,
   'videoFilters': instance.videoFilters
       .map((e) => _$VideoFilterEnumMap[e]!)
       .toList(),
@@ -239,7 +235,7 @@ final class SettingsControllerProvider
 }
 
 String _$settingsControllerHash() =>
-    r'4514df87d1ffc8913f1e35b30167490bd1bbc386';
+    r'9a7d402cb9889d3bd700820c24f4d0bf42356bc4';
 
 abstract class _$SettingsController extends $Notifier<Settings> {
   Settings build();